+++ lib/src/_core/dependencies/deps_container.dart
@@ -6,11 +6,13 @@ import 'package:tracksu/src/_core/l10n/locale_controller.dart';
 import 'package:tracksu/src/session/session_controller.dart';
 import 'package:tracksu_network/tracksu_network.dart';
 import 'package:tracksu_storage/tracksu_storage.dart';
+import 'package:tracksu/src/_shared/content/content_media_controller.dart';
 
 final class DepsContainer {
   const DepsContainer({
     required this.appRouter,
     required this.localeController,
+    required this.contentMediaController,
     required this.authRepository,
     required this.oauthClientCredentials,
     required this.oauthCallbackLinkSource,
@@ -24,6 +26,7 @@ final class DepsContainer {
 
   final TracksuAppRouter appRouter;
   final LocaleController localeController;
+  final ContentMediaController contentMediaController;
   final AuthRepository authRepository;
   final OAuthClientCredentials oauthClientCredentials;
   final OAuthCallbackLinkSource oauthCallbackLinkSource;
@@ -38,6 +41,7 @@ final class DepsContainer {
     appRouter.dispose();
     sessionController.dispose();
     localeController.dispose();
+    contentMediaController.dispose();
     restClient.close();
     publicRestClient.close();
     oauthRestClient.close();