+++ packages/tracksu_storage/lib/tracksu_storage.dart
@@ -1,5 +1,7 @@
 library;
 
+export 'src/content_media_store.dart';
+
 export 'src/flutter_secure_locale_store.dart';
 export 'src/flutter_secure_token_store.dart';
 export 'src/flutter_secure_oauth_transaction_store.dart';